+++ fixtures/esm/forge-dependency-js-config/forge.config.js
@@ -1,4 +1,4 @@
-module.exports = {
+export default {
   packagerConfig: {
     name: 'forge-dependency-js-config',
   },
