+++ fixtures/esm/forge-dependency-linked-js-config/custom-config.js
@@ -1,4 +1,4 @@
-module.exports = {
+export default {
   packagerConfig: {
     name: 'forge-dependency-linked-js-config',
   },
