+++ packages/@wdio_electron-utils/src/config/forge.ts
@@ -1,6 +1,6 @@
 import path from 'node:path';
-import { pathToFileURL } from 'node:url';
 
+import { readConfig } from './read';
 import log from '../log.js';
 import { APP_NAME_DETECTION_ERROR } from '../constants.js';
 
@@ -34,10 +34,11 @@ export async function getConfig(pkg: NormalizedReadResult): Promise<ForgeBuildIn
 
   if (!forgePackageJsonConfig || isConfigFilePath) {
     // if no forge config or a linked file is found in the package.json, attempt to read Forge JS-based config
-    const forgeConfigPath = path.join(rootDir, isConfigFilePath ? forgePackageJsonConfig : 'forge.config.js');
+    const forgeConfigFileName = isConfigFilePath ? forgePackageJsonConfig : 'forge.config.js';
+    const forgeConfigPath = path.join(rootDir, forgeConfigFileName);
     try {
       log.info(`Reading Forge config file: ${forgeConfigPath}...`);
-      forgeConfig = ((await import(pathToFileURL(forgeConfigPath).toString())) as { default: ForgeConfig }).default;
+      forgeConfig = (await readConfig(forgeConfigFileName, rootDir)).result as ForgeConfig;
     } catch (_) {
       log.warn(`Forge config file not found or invalid at ${forgeConfigPath}.`);
       return undefined;
